filter all chirps

+++ sql/queries/chirps.sql
@@ -12,8 +12,12 @@ RETURNING *;
 -- name: ResetChirps :exec
 TRUNCATE TABLE chirps CASCADE;
 
+-- sqlc dynamic queries workaround: https://dizzy.zone/2024/07/03/SQLC-dynamic-queries/
 -- name: GetAllChirps :many
-SELECT * FROM chirps ORDER BY created_at ASC;
+SELECT * FROM chirps
+WHERE id IS NOT NULL
+AND (NOT @has_user_id::boolean OR user_id = $1)
+ORDER BY created_at ASC;
 
 -- name: GetChirpById :one
 SELECT * FROM chirps WHERE id = $1;
